// Entorno de pruebas (funcional)/Frontend NO BORRAR/DiabetesNoteBook/src/app/interfaces/usuario.interface.ts
import { Sexo } from 'src/app/enums/register.enum';

export interface IUsuarioUpdate {
  id?: number;
  avatar: string;
  userName: string;
  nombre: string;
  primerApellido: string;
  segundoApellido: string;
  email: string;
  sexo: Sexo;
  edad: number;
  peso: number;
  altura: number;
  actividad: string;
  tipoDiabetes: string;
  medicacion: string[];
  insulina: boolean;
  usuarioMedicacions?: any[];
}